requirements.txt
flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0
langchain==0.3.17  # Replace with the actual version you are using
flask-session==0.4.0
PyPDF2==3.0.1
tiktoken==0.8.0